fixed interface ; bug added for colly in start.go
